pyproject.toml
[project]
name = "blockchainchess"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chess>=1.11.2",
    "eth-account>=0.13.7",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.8",
]
